Refactor the right sidebar view
The way the old right sidebar was structured made it very difficult for the chat input
to communicate with the rest of the view. There were lots of nested and subnested
components, which made passing props and refs confusing.

I've refactored the right column and nested components specific to that column.

In addition, this commit cleans up some linting errors, makes the emoji picker insert
the emoji into the text input (rather than sending a standalone message) and handles
scrolling to the bottom whenever the current user sends a message, no matter where
they are in the story.