Don't delete channel or community relationships
Rather than deleting relationships that a user has to a community or channel, we can instead keep the record but set isMember to false and turn off notifications.

This will be helpful when dealing with notification logic down the road - e.g. if a user deletes a community, it will automatically turn off all community + channel notifications for all users in that community!