Fix notifications listening
Previously we were listening to new documents in `notifications` and changes to the `seen` or `read` fields in the `usersNotifications` table for our subscriptions. This meant whenever a new notification was added or it's seen or read status changed we'd send the changes down to the client.

While this works, it's suboptimal for browser push notifications since we don't want to send a push notification when you look at your notifications :sweat_face: What we really want is to listen to new documents in `notifications` _and_ to changes to the `usersNotifions` table but _only_ if the changes come from `athena`. We don't want to send down the changes if a user marks a notification as seen/read.

I introduce a new field on `usersNotifications` to solve this, a simple `entityAddedAt`. It's set in athena whenever a notification in the same context comes in, meaning we send way less data down to the client and browser push notifications work as expected! :tada:

/cc @brianlovin I'm tagging you here so you'll get this commit explicitly in your notifications and can see the reasoning for this change.