Stop redirect from render on dashboard
This was breaking the app for me if I ended up having an invalid sesion.
(which happens every so often when we deploy) It made the app redirect
forever and ever until I manually cleared out the data from local
storage, thusly signing me out-this is obviously not something users
will do.

This fixes it by moving the loading state to the correct location in the
dashboard and stopping to redirect in the render method.