Select "General" in the thread composer
Right now, all communities have multiple default channels. This means in
the threadComposer we just show the first (by creation date?) channel
that is a default one.

This is super annoying because lots of users post in the wrong channels.
See this thread (and the whole channel) for an example:
https://spectrum.chat/styled-components/stylelint-processor?thread=05a3c7fb-6a95-4d61-bef2-57d82ce34b90

This patch fixes it by selecting the channel with the "general" slug if
there is one, even if there are multiple default channels.