Downgrade to graphql@0.10.0
I hit a wall here unfortunately—upgrading to graphql@0.11 would require
an upgrade of subscriptions-transport-ws@0.9. The issue is that
subscriptions-transport-ws@0.9 deprecates the subscription API we use on
the client (that's the warning you see in the console in dev mode), but
the replacement for the hybrid ws/http transport is a) undocumented and
b) it would require an update to Apollo client 2.0, which is a much
bigger task than just upgrading a couple packages. (will tackle that
sometime very soon though, since we need it for the mobile apps anyway)

Assuming tests pass now this should be good to go.