Billing route logic
This requires a community to have an administrator email on file in order to manage billing settings.
If they dont have an admin email on file, they will be prompted to enter one, which then sends
a confirmation email and goes through the whole pending email flow that is very similar to how we confirm
user email addresses.

If an admin email is on file, the user can now see a card form powered by Stripe in the billing tab...more to come here