Do not call `markSingleNotificationSeen` for all client-side notifications
As you can see in the diff, we were calling the `markSingleNotificationSeen` mutation from the client for all notifications when clicking "Mark all as seen", even though we also call the `markAllAsSeen` mutation.

This results in a lot of unnecessary work and huge spikes in response times.

This fixes it by not unnecessarily calling `markSingleNotificationSeen`.