Temporarily log all post request paths
For some reason we see a ton of POST requests that do not go to our API.
I want to figure out what those are, as they get rate limited fairly
frequently.

This adds a temporary `console.log` for every POST request that does not
hit /api. We should deploy this to production, monitor for a couple
minutes to figure out what those POST requests are, then revert.