Fetch community metadata for user community connection
We used to not fetch the community metadata for the user community
connection because we did not display any of it unless you visited a
community page, which was not very common, and it was quite expensive to
calculate.

Now, we actually do display this everywhere and it is cached for a long
time so it's not expensive to fetch. By also fetching this from the
community connection, the frontend won't do another request to load the
community data when switching between communities, heavily reducing the
number of API requests we are seeing.